requirements.txt
flask>=3.0.0
pefile>=2023.2.7
capstone>=5.0.1
python-magic-bin>=0.4.14
yara-python>=4.3.1
requests>=2.31.0
jinja2>=3.1.2
rich>=13.7.0
click>=8.1.7
pyyaml>=6.0.1
mcp[cli]>=1.0.0
psutil>=5.9.8
colorama>=0.4.6
tabulate>=0.9.0
reportlab>=4.0.0
python-docx>=1.0.0
# New dependencies added in v2
oletools>=0.60        # OLE/macro analysis
pdfminer.six>=20221105 # PDF text extraction
pyelftools>=0.31      # ELF binary parsing
speakeasy-emulator>=2.0.1  # CPU emulation (Mandiant) — pip install speakeasy-emulator
